refactor: move updateHits to actor method
